+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added backup and restore methods to `Connection` (Issue #46)
 - Added `Connection::$backupOnMigration` option, defaults to `true`
 - Added `Dsn` helper class to parse DSN strings
+- Changed `FileHelper::generateRandomFilename()` signature
 - Locked database schema to MySql
 
 ## 2.5.8 (Sep 23, 2025)